feat: create team pages and put team list in navigator
