+++ app/[teamId]/components/buttons/acceptInvitationButton.tsx
