+++ app/[teamId]/components/buttons/leaveTeamButton.tsx
