+++ app/[teamId]/components/teamDetails.tsx
