+++ app/[teamId]/components/teamFooter.tsx
