+++ app/[teamId]/components/teamHeader.tsx
@@ -4,8 +4,6 @@ type TeamHeaderProps = {
 
 export function TeamHeader({ teamName }: TeamHeaderProps) {
   return (
-    <div className="flex items-center justify-between">
-      <h2 className="text-2xl font-bold">{teamName}</h2>
-    </div>
+    <h1 className="text-3xl font-bold">{teamName}</h1>
   )
 }
