+++ app/components/navigator.tsx
@@ -1,19 +1,25 @@
-'use client'
-
 import { EyeIcon, HomeIcon, PlusIcon, UsersIcon } from '@/components/icons'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 
-export default function Navigator() {
-  const pathname = usePathname()
+import { getTeamsByUserId } from '@/lib/api/team'
+import { getServerSession } from 'next-auth'
+import { getInvitedTeams } from '@/lib/api/team'
 
-  const getLinkClass = (path: string) => {
-    return pathname === path
-      ? 'font-medium text-bold bg-light-blue scale-105'
-      : 'transition ease-in-out text-muted-foreground hover:bg-light-blue hover:scale-105 hover:duration-300'
-  }
+export default async function Navigator() {
 
   const commonStyle = 'flex items-center gap-3 p-2 rounded-lg'
+
+  const session = await getServerSession()
+
+  let teams = []
+  let invitedTeams = []
+  if (session) {
+    const teamsRes = await getTeamsByUserId()
+    teams = teamsRes.teams
+    const invitedTeamsRes = await getInvitedTeams()
+    invitedTeams = invitedTeamsRes.teams
+  }
 
   return (
     <nav className="flex w-1/6 flex-col gap-2 bg-slate-100 p-3 shadow-md">
@@ -21,18 +27,17 @@ export default function Navigator() {
         <HomeIcon className="h-6 w-6" />
         Home
       </Link> */}
-      <Link href="/team" className={commonStyle + ' ' + getLinkClass('/team')}>
-        <UsersIcon className="h-6 w-6" />
-        My Teams
-      </Link>
-      <Link href="/record" className={commonStyle + ' ' + getLinkClass('/record')}>
-        <PlusIcon className="h-6 w-6" />
-        Create Record
-      </Link>
-      <Link href="/view" className={commonStyle + ' ' + getLinkClass('/view')}>
-        <EyeIcon className="h-6 w-6" />
-        View Records
-      </Link>
+      <UsersIcon className="h-6 w-6" />My Teams
+      {teams?.map((team) => (
+        <Link
+          href={`/${team.id}`}
+          key={team.id}
+        >
+          {team.name}
+        </Link>
+      ))
+      }
     </nav>
   )
 }
+
